reduce runtime dependencies using `core-js`.
This replaces nearly all the runtime dependencies with `core-js` equivalents.
This should maintain support for old browsers, while drastically reducing the size.
This commit drops the browserified size down to just over 600 lines.
It was 1200 lines previously (and over 8,000 in the empower module).

The extra good news for AVA, is that babel-runtime pulls in `core-js`
anyways, so our savings should be even more substantial.