add ability to track non enhanced methods.
This adds the ability to track methods that, for whatever reason,
are not good targets for runtime enhancement. It is included to
provide an easy means for test runners to track their assertion plans
(i.e. `tap.plan(3)`). It is able to use the call signatrue to extract
the `message` property before calling the assertion.

This provides implementors of test runners a more unified API for
inspecting assertions, whether they are enhanced or not. While this
is a little out of scope for power-assert, I believe it is worth
including.